requirements.txt
streamlit
opencv-python
sympy
networkx
pandas
pillow
streamlit-webrtc
matplotlib